refactor: SessionDto Controller/Service 레이어 분리
