+++ src/main/kotlin/nexters/admin/service/session/SessionService.kt
@@ -1,7 +1,8 @@
 package nexters.admin.service.session
 
+import nexters.admin.controller.session.CreateSessionRequest
+import nexters.admin.controller.session.UpdateSessionRequest
 import nexters.admin.domain.session.Session
-import nexters.admin.exception.BadRequestException
 import nexters.admin.exception.NotFoundException
 import nexters.admin.repository.SessionRepository
 import org.springframework.data.repository.findByIdOrNull
@@ -38,7 +39,8 @@ class SessionService(
     }
 
     fun updateSession(sessionId: Long, request: UpdateSessionRequest) {
-        val session = sessionRepository.findByIdOrNull(sessionId) ?: throw NotFoundException.sessionNotFound()
+        val session = sessionRepository.findByIdOrNull(sessionId)
+                ?: throw NotFoundException.sessionNotFound()
 
         session.apply {
             title = request.title
